refactor: consolidate main module
