+++ run.py
@@ -1,6 +1,6 @@
 # Entry-point alias so tests can import "run"
-# AI-AGENT-REF: delegate entrypoint to package __main__
-from ai_trading.__main__ import main
+# AI-AGENT-REF: delegate entrypoint to canonical main module
+from ai_trading.main import main
 
 if __name__ == "__main__":
     main()
