+++ tests/test_additional_coverage.py
@@ -16,7 +16,7 @@ import config
 import meta_learning
 import ml_model
 import risk_engine
-import ai_trading.__main__ as main
+import ai_trading.main as main
 import utils
 import validate_env
 from strategies.mean_reversion import MeanReversionStrategy
@@ -69,9 +69,9 @@ def test_create_flask_routes():
     sys.modules['flask'] = flask_mod
     sys.modules['flask.testing'] = types.ModuleType('flask.testing')
     sys.modules['flask.testing'].FlaskClient = DummyClient
-    sys.modules.pop('ai_trading.__main__', None)
+    sys.modules.pop('ai_trading.main', None)
     import importlib
-    main_mod = importlib.import_module('ai_trading.__main__')
+    main_mod = importlib.import_module('ai_trading.main')
     app = main_mod.create_flask_app()
     client = app.test_client()
     assert client.get("/health").json() == {"status": "ok"}
